eth indexer: refactor helios and direct rpc, fix sign bidirectional flow (#600)
* refactor eth indexer so helios and direct rpc use mostly same code path

* upgrade helios version and simplify rpc call

* fix from_address issue

* set optimistic_requests to true for integration tests to run

---------
