+++ chain-signatures/node/src/indexer_eth/indexer_eth_helios.rs
@@ -1,0 +1,171 @@
+use crate::indexer_eth::EthConfig;
+use alloy::eips::{BlockId, BlockNumberOrTag};
+use alloy::primitives::Address;
+use alloy::primitives::Bytes;
+use alloy::rpc::types::TransactionRequest;
+use helios::ethereum::{config::networks::Network, EthereumClient, EthereumClientBuilder};
+use std::path::PathBuf;
+use std::str::FromStr;
+use std::sync::Arc;
+use tokio::time::Duration;
+
+#[derive(Clone)]
+pub struct HeliosEthereumClient {
+    client: Arc<EthereumClient>,
+    max_retries: u8,
+    base_delay: Duration,
+}
+
+impl HeliosEthereumClient {
+    fn new(client: EthereumClient, max_retries: u8, base_delay: Duration) -> Self {
+        Self {
+            client: Arc::new(client),
+            max_retries,
+            base_delay,
+        }
+    }
+
+    pub async fn get_block(
+        &self,
+        block_id: alloy::rpc::types::BlockId,
+    ) -> Option<alloy::rpc::types::Block> {
+        self.fetch_block(block_id).await
+    }
+
+    pub async fn get_block_receipts(
+        &self,
+        block_id: alloy::rpc::types::BlockId,
+    ) -> anyhow::Result<Option<Vec<alloy::rpc::types::TransactionReceipt>>> {
+        self.client
+            .get_block_receipts(block_id)
+            .await
+            .map_err(|err| anyhow::anyhow!("Failed to get block receipts for block: {:?}", err))
+    }
+
+    pub async fn get_nonce(
+        &self,
+        address: Address,
+        block_id: alloy::rpc::types::BlockId,
+    ) -> anyhow::Result<u64> {
+        self.client
+            .get_nonce(address, block_id)
+            .await
+            .map_err(|err| {
+                anyhow::anyhow!(
+                    "Failed to get nonce for address {address:?} and block id {block_id:?}: {:?}",
+                    err
+                )
+            })
+    }
+
+    pub async fn get_transaction_by_hash(
+        &self,
+        tx_hash: alloy::primitives::B256,
+    ) -> anyhow::Result<Option<alloy::rpc::types::Transaction>> {
+        self.client.get_transaction(tx_hash).await.map_err(|err| {
+            anyhow::anyhow!("Failed to get transaction by hash {tx_hash:?}: {:?}", err)
+        })
+    }
+
+    pub async fn call(
+        &self,
+        from: Address,
+        to: Address,
+        data: Bytes,
+        block_number: u64,
+    ) -> anyhow::Result<Bytes> {
+        // Build a base tx *without* the sentinel max gas
+        let mut tx = TransactionRequest::default()
+            .from(from)
+            .to(to)
+            .input(alloy::rpc::types::TransactionInput::both(data.clone()));
+
+        // 1) Estimate
+        let est = self
+            .client
+            .estimate_gas(&tx, BlockId::Number(BlockNumberOrTag::Number(block_number)))
+            .await
+            .unwrap_or(3_000_000u64); // fallback
+
+        // 2) Add 20% buffer, but keep < 16,777,216
+        let mut gas = (est as f64 * 1.2) as u64;
+        if gas > 16_777_216 {
+            gas = 16_777_216;
+        }
+
+        // 3) Apply gas limit
+        tx = tx.gas_limit(gas);
+
+        // 4) Execute the call
+        self.client
+            .call(&tx, BlockId::Number(BlockNumberOrTag::Number(block_number)))
+            .await
+            .map_err(|err| anyhow::anyhow!("Failed to call: {err:?}"))
+    }
+
+    pub async fn get_latest_block_number(&self) -> anyhow::Result<u64> {
+        let Some(block) = self
+            .fetch_block(BlockId::Number(BlockNumberOrTag::Latest))
+            .await
+        else {
+            return Err(anyhow::anyhow!("Latest block not found"));
+        };
+        Ok(block.header.number)
+    }
+
+    // retry getting block from helios with exponential backoff
+    async fn fetch_block(&self, block_id: BlockId) -> Option<alloy::rpc::types::Block> {
+        let mut retries = 0;
+        loop {
+            match self.client.get_block(block_id, false).await {
+                Ok(Some(block)) => return Some(block),
+                Ok(None) => {
+                    tracing::warn!("Block {block_id} not found from Helios client");
+                    return None;
+                }
+                Err(e) => {
+                    if retries < self.max_retries {
+                        retries += 1;
+                        let delay = self.base_delay * 2u32.pow((retries - 1) as u32);
+                        tracing::warn!(
+                        "Failed to fetch block number {block_id} from Helios client: {e:?}, retrying"
+                    );
+                        tokio::time::sleep(delay).await;
+                        continue;
+                    }
+                    tracing::warn!(
+                    "Failed to fetch block number {block_id} from Helios client: {e:?}, exceeded maximum retry"
+                );
+                    return None;
+                }
+            }
+        }
+    }
+}
+
+pub async fn build_client(eth: EthConfig) -> anyhow::Result<HeliosEthereumClient> {
+    let Ok(network) = Network::from_str(eth.network.as_str()) else {
+        return Err(anyhow::anyhow!("Network input incorrect: {}", eth.network));
+    };
+    let client = EthereumClientBuilder::new()
+        .network(network)
+        .consensus_rpc(&eth.consensus_rpc_http_url)
+        .map_err(|err| anyhow::anyhow!("failed to build consensus rpc: {err:?}"))?
+        .execution_rpc(&eth.execution_rpc_http_url)
+        .map_err(|err| anyhow::anyhow!("failed to build execution rpc: {err:?}"))?
+        .data_dir(PathBuf::from(&eth.helios_data_path))
+        .with_file_db()
+        .build()
+        .map_err(|err| anyhow::anyhow!("failed to build helios client: {err:?}"))?;
+    tracing::info!("Built Helios client on network {}", network);
+    client
+        .wait_synced()
+        .await
+        .map_err(|err| anyhow::anyhow!("Failed to wait for synced: {err:?}"))?;
+
+    Ok(HeliosEthereumClient::new(
+        client,
+        6,
+        Duration::from_millis(200),
+    ))
+}
